separate source into source and destination. fix issue with serialization
