+++ src/Main.scala
@@ -1,6 +1,6 @@
-import flowbit.{FlowBitImpl, TestDestination, TestSource}
+import flowbit.endpoints.{TestDestination, TestSource}
+import flowbit.FlowBitImpl
 import org.apache.kafka.streams.KeyValue
-import org.apache.kafka.streams.kstream.Predicate
 
 object Main {
 
@@ -23,7 +23,7 @@ object Main {
 
     println("adding filter")
     flowbit.addFilter[String, String]("filter1", "toBeFiltered", List("toBeMapped"),
-      new TestPred())
+      (k, v) => k.last.toInt % 2 == 0)
 
     println("adding map")
     flowbit.addMap[String, String, String, String]("map1", "toBeMapped", List("done"),
@@ -34,10 +34,4 @@ object Main {
     flowbit.addConsumer[String, String]("consumer1", "done", "group1", dest)
   }
 
-}
-
-class TestPred extends Predicate[String, String] {
-  override def test(k: String, v: String): Boolean = {
-    k.last.toInt % 2 == 0
-  }
 }
