+++ src/flowbit/FilterComponent.scala
@@ -1,5 +1,6 @@
 package flowbit
 
+import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.{KafkaStreams, StreamsBuilder, StreamsConfig}
 import org.apache.kafka.streams.kstream.{KStream, Predicate}
 
@@ -10,8 +11,13 @@ final class FilterComponent[A,B](id: String, server: String, fromTopic: String,
   // Filter-specific configs
   properties.put(StreamsConfig.APPLICATION_ID_CONFIG, id)
 
-  override def execute(): Unit = {
+  // Add to stop casting and serialization issues with strings
+  val strings = Serdes.String()
+  properties.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, strings.getClass().getName())
+  properties.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, strings.getClass().getName())
 
+
+  override def execute(): Unit = {
     val builder: StreamsBuilder = new StreamsBuilder()
     val stream: KStream[A, B] = builder.stream[A, B](fromTopic)
     val filteredStream: KStream[A, B] = stream.filter(pred)
