+++ src/flowbit/MapComponent.scala
@@ -1,8 +1,8 @@
 package flowbit
 
+import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.kstream.{KStream, KeyValueMapper}
 import org.apache.kafka.streams.{KafkaStreams, KeyValue, StreamsBuilder, StreamsConfig}
-import org.apache.kafka.streams.scala.ImplicitConversions._
 
 
 final class MapComponent[A,B,C,D](id: String, server: String, fromTopic: String,
@@ -10,6 +10,11 @@ final class MapComponent[A,B,C,D](id: String, server: String, fromTopic: String,
 
   // Map-specific configs
   properties.put(StreamsConfig.APPLICATION_ID_CONFIG, id)
+
+  // Add to stop casting and serialization issues with strings
+  val strings = Serdes.String()
+  properties.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, strings.getClass().getName())
+  properties.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, strings.getClass().getName())
 
   override def execute(): Unit = {
